# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Encryptr Server API",
    "description": "Encrypted AI conversation backend. Message content is always client-side AEAD ciphertext; the server stores ciphertext, nonce, and auth tag but can never decrypt it. The `/chat/stream` WebSocket provides end-to-end encrypted streaming inference.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "paths": {
    "/chat/stream": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Establish the end-to-end encrypted streaming chat WebSocket.",
        "description": "This is a WebSocket upgrade, not a normal request/response. After upgrade the\nclient and server perform an X25519 handshake, then exchange AES-256-GCM\nencrypted JSON frames (`prompt` → `chunk*` → `done`). See the module docs for\nthe frame protocol. OpenAPI cannot model the frame exchange itself.",
        "operationId": "chat_stream",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Access JWT — passed as a query parameter because browsers cannot set\nthe `Authorization` header on a WebSocket upgrade.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols — WebSocket upgrade"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/keys": {
      "get": {
        "tags": [
          "keys"
        ],
        "summary": "List the caller's public keys.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Public keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "Upload a device public-key pair (Ed25519 + X25519).",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Stored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate label",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "The error envelope returned by every endpoint on failure. Mirrors\n`crate::error::AppError`'s HTTP body (defined here so it can carry a\n`ToSchema` derive without leaking into the error type).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `validation_error`, `unauthorized`,\n`not_found`, `conflict`, `rate_limited`, `provider_error`,\n`internal_error`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, safe-to-display message."
          }
        }
      },
...
      "CreateKeyRequest": {
        "type": "object",
        "required": [
          "ed25519_public_key",
          "x25519_public_key"
        ],
        "properties": {
          "ed25519_public_key": {
            "type": "string",
            "description": "Base64-encoded 32-byte Ed25519 public key."
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "x25519_public_key": {
            "type": "string",
            "description": "Base64-encoded 32-byte X25519 public key."
          }
        }
      },
...
      "KeyResponse": {
        "type": "object",
        "required": [
          "id",
          "ed25519_public_key",
          "x25519_public_key",
          "label",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "ed25519_public_key": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "x25519_public_key": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from `POST /login`."
      }
    }
  },
...
}